requirements.txt
# Core Django
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
django-cacheops==7.0.2
django-debug-toolbar==4.2.0

# Environment configuration
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Caching & Database
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
dj-database-url==2.1.0

# Storage
boto3==1.34.0
pillow==10.2.0

# AI services
openai==1.12.0
anthropic==0.7.7
langchain==0.1.0
transformers==4.37.2

# Utils
requests==2.31.0
pydantic==2.5.3
python-dateutil==2.8.2
markdown==3.5
bleach==6.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
